Add extra parameter to get store for reducer function and middleware
