+++ src/index.js
@@ -1,70 +1,47 @@
-const { hydrate, getStore } = require('./clientSide.js');
-const { init, serve } = require('./serverSide.js');
+const React = require('react');
+const ReactDOMServer = require('react-dom/server');
+const { StaticRouter } = require('react-router-dom');
+const { createStore } = require('redux');
+const { Provider } = require('react-redux');
+const fs = require('fs');
 
-console.log("hyd>>>>>>..", hydrate);
+const config = {};
+const init = (inputs) => {
+  config.html = inputs.html;
+  config.App = inputs.App;
+  config.reducer = inputs.reducer;
+  config.id = inputs.id;
+};
+
+const serve = (req, res) => {
+  const { App, reducer } = config;
+
+  const context = {};
+  const serveStore = createStore(reducer);
+  const preloadedState = serveStore.getState();
+
+  const stringComponent = ReactDOMServer.renderToString(
+    <Provider store={serveStore}>
+      <StaticRouter location={req.url} context={context}>
+        <App />
+      </StaticRouter>
+    </Provider>
+  );
+  if (context.url) {
+    res.status = 302;
+    res.redirect(context.url);
+  } else {
+    fs.readFile(config.html, 'utf8', (err, data) => {
+      if (err) throw err;
+      const regEx = new RegExp(`<div id="${config.id}"></div>`, 'gi');
+      const document = data.replace(regEx, `<div id="${config.id}">${stringComponent}</div><script>window.__HELIUM_CONFIG__ = ${JSON.stringify(preloadedState).replace(/</g, '\\u003c')}</script>`);
+      res.write(document);
+      res.end();
+    });
+  }
+};
 
 module.exports = {
-  hydrate,
-  getStore,
   init,
   serve,
 };
-
-// let preloadedState = {};
-
-// const getStore = () => {
-//   console.log(preloadedState);
-//   console.log(config.reducer);
-//   createStore(config.reducer, preloadedState);
-// };
-
-// const init = (inputs) => {
-//   console.log('inside init >>>>>>>>>>>>>>>');
-//   config.html = inputs.html;
-//   config.App = inputs.App;
-//   config.reducer = inputs.reducer;
-
-//   config.renderStore = createStore(config.reducer);
-//   preloadedState = config.renderStore.getState();
-// };
-
-// const serve = (req, res) => {
-//   const fs = require('fs');
-//   console.log('inside serve >>>>>>>>>>>>>>>');
-//   console.log(config);
-//   const { App, renderStore } = config;
-  
-//   const context = {};
-//   const stringComponent = ReactDOMServer.renderToString(
-//     <Provider store={renderStore}>
-//       <StaticRouter location={req.url} context={context}>
-//         <App />
-//       </StaticRouter>
-//     </Provider>
-//   );
-//   if (context.url) {
-//     res.status = 302;
-//     res.redirect(context.url);
-//   } else {
-//     fs.readFile(config.html, 'utf8', (err, data) => {
-//       if (err) throw err;
-//       const regEx = new RegExp(`<div id="${process.env.ID}"><\/div>`, 'gi');
-//       const document = data.replace(regEx, `<div id="${process.env.ID}">${stringComponent}</div>`);
-//       res.write(document);
-//       res.end();
-//     });
-//   }
-// };
-
-// const hydrate = (e, inputId) => {
-//   // process.env.ID = inputId;
-//   console.log("iddddddd", inputId);
-//   ReactDOM.hydrate(e, document.getElementById(inputId));
-// };
-
-// export default hydrate;
-// exports = {
-//   getStore,
-//   init,
-//   serve,
-// };
